cmake: Exclude host directories from search
This is a follow-up to b65c24db ("Add NDK toolchain's sysroot to search paths").
With `CMAKE_FIND_ROOT_PATH_MODE` variables set to `BOTH` or `NEVER`, CMake still
includes system directories in library search paths:

  -- Found LibBfd: /usr/lib64/libbfd.a
  -- Found LibOpcodes: /usr/lib64/libopcodes.a

Using `ONLY`[1] prevents this, and is consistent with the NDK toolchain[2]. Note
that this also requires our output directory to be added to root paths.

[1] https://cmake.org/cmake/help/latest/variable/CMAKE_FIND_ROOT_PATH_MODE_LIBRARY.html
[2] https://github.com/Kitware/CMake/blob/v3.22.2/Modules/Platform/Android-Initialize.cmake#L48